[composebox][omnnibox] Add page classifications for omnibox composebox.
Previously, the omnibox popup composebox was using the NTP_COMPOSEBOX
page classification as a placeholder to test receiving suggestions. This
change rolls over the proto changes which added the 3 relevant omnibox
composebox page classifications (NTP, SRP, OTHER) and pipes it into the
new OmniboxPopupComposeboxClient. From the user perspective, this should
not change anything and is strictly meant for logging.

The logic for assigning the NTP, SRP, OTHER page classifications is the
same as what the omnibox currently uses (with OTHER covering a couple
more cases).

This also creates a new OmniboxComposeboxHandler which extends the
ComposeboxHandler and allows us to pass in an OmniboxController with the
correct OmniboxClient.

Change-Id: I88dcfc9e8f7d32a7aa05f899dc039792984e74b8
Bug: 458054953
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7147009
Auto-Submit: Nihar Majmudar <niharm@google.com>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Commit-Queue: Nihar Majmudar <niharm@google.com>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1545411}