[Omnibox WebUI] Fix missing WebUI popup
RoundedOmniboxResultsFrame clamped the frame height to 0 when the WebUI
requested a minimum size (1px) for 0 results. This zero-height bounds
caused the WebContent to be hidden, stopping the renderer from
generating and sending new resize events.

Fix is clamp the bounds to a centered 1x1 pixel instead of 0 height.
This will keep the widget alive but also keep it unnoticeable to the
user.

This is a short-term workaround. There will be a separate bug to tackle
refactoring the regular popup so it does not send min-height requests.

Bug: 460332162, b/460908495
Change-Id: I22c035ed117540fadccaf50c73193ab3ef368add
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7159529
Reviewed-by: Marlon Facey <mfacey@chromium.org>
Commit-Queue: David Yeung <dayeung@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1545480}