[CC Clip Paths] Use paint property paint clip rect as dst rect
This CL is the 2nd of a series of CLs that will allow for the paint
rect of a cc clip path animation to be an arbitrary value, which would
fix multiple perf issues related to CC clip paths.

In this change, the clip path bounding rect is no longer re-computed
in-situ in ClipPathClipper, but re-used from the paint properties.
This ensures that any calculation of the rect only needs to happen
once, which is important for cc clip path animations as calculating a
minimal bounding rect requires walking all keyframes and extrapolating
paths, which is an expensive operation.

As a side effect, this also ensures all translations applied to the
paint clip rect are also applied to the paint worklet deferred image.
To prevent the paint offset from being applied twice, the translation
in ClipPathClipper::PaintClipPathAsMaskImage is removed for cc clip
paths. Additionally, this change also removes adjustment of the
reference_box and instead adjusts generated paths by the paint offset,
as is done for main thread clip paths. This part of the change is not
strictly necessary, but is done here because that adjustment of the
reference_box produces numerous subpixel differences caused by
floating-point imprecision. This causes ref tests to fail and require
fuzziness to be added. Doing this change now avoids that, as well as
being a step towards fixing crbug.com/459701868.

This CL should have no functional changes.

Bug: 447935064, 449779020

Change-Id: Ifbbd19165503ffa63e1ddcf9205721b736e7f4d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7145798
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: Claire Chambers <clchambers@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1545527}