WebView Web Performance Metrics API: Consolidate listener/notification code
This cl creates a new class, AwNavigationClient, to handle all of the code for keeping track of
AwNavigationListeners and forwarding notifications onwards to them. The list
of listeners is now contained in the new class and all
modifications to the list will take place in this one central class.

This work will make adding new performance api methods easier and cleaner as
the code will all be in one centralised place.

Bug: b:432696062
Change-Id: I3839a1ae8559edd70ed4632997cf44f97e530737
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7141940
Commit-Queue: Joanne de Abreu <jdeabreu@google.com>
Reviewed-by: Peter Conn <peconn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1545769}