[Blink] Remove FlushReason::kNone
Neither FlushReason::kNone nor FlushReason::kOther are ever referenced
other than being passed [1]. Hence, we can replace usage of the former
with usage of the latter.

[1] There is a minor exception, which is a DCHECK in
CanvasResourceProvider that the flush reason is not kNone if there are
releasable draw ops. This DCHECK is not important to preserve and this
CL removes it.

Bug: 352263194
Change-Id: I01aae2e1b5543e43d19c23aa8c9424b2d16df2c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7157032
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1545772}