cros-geo-caching: Rename geolocation_handler to geolocation_handler_impl
This class serves as a SystemLocationProvider delegate, which was
implemented as a concrete class without an interface. In the subsequent
CL we'll extract the GeolocationHandler interface from the
implementation.

Rename the "geolocation_handler*" files to "*_impl", we'll introduce the
geolocation_handler.h as an interface separately in the subsequent CL.
Doing it altogether confuses Git/Gerrit on what file actions has been
made. See this discussion[1] for details.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/7088820/comment/71a5ecd6_1cd4f8a8/

Bug: 455571169
Change-Id: I973a8254e07eb15513f5490d067f0d8c09989a36
Bypass-Check-License: Renamed files
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7148482
Auto-Submit: Zauri Meshveliani <zauri@chromium.org>
Commit-Queue: Zauri Meshveliani <zauri@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Alvin Ji <alvinji@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1545781}