Remove sentinel values in std::arrays in trace_event_args_allowlist.cc
Even though https://crrev.com/1404094 rewrote the data in this file to
use std::array, it didn't take advantage of std::array::size(). Do that
in this CL and remove the sentinel values. Then rewrite the the
for-loops that iterate through the arrays to be range-based. Also mark
one of the arrays as constexpr and remember to include <array>.

Change-Id: Ie0f5dc4e8feba87ed5c8454f924f588442685762
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7159299
Reviewed-by: Mikhail Khokhlov <khokhlov@google.com>
Auto-Submit: Lei Zhang <thestig@chromium.org>
Commit-Queue: Mikhail Khokhlov <khokhlov@google.com>
Cr-Commit-Position: refs/heads/main@{#1545821}