[Print] Align PDF default print scale to HTML print
The default PDF print scale was CUSTOMISED 100, which does not scale or
centre the content. This could cause content to be cropped in the print
preview if it extended beyond the paper size. To address this, this CL
aligns the default scale of PDF print to HTML print, which is
`kCenterShrinkToFitPaper`, which shrinks the overflow content and centre
the smaller content.

This change is controlled by flag kAlignPdfDefaultPrintSettingsWithHTML.

Bug: 402056134
Change-Id: If891175d3dd1ea2a5f1fecedb132c4ab2f5c0d58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6703647
Reviewed-by: Andy Phan <andyphan@chromium.org>
Commit-Queue: cathie chen <cchen@igalia.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1545831}