Add ChromeOS device login screen CryptographyCompliance policies
This CL adds device-level equivalents for the CryptographyCompliance
policies PreferSlowKexAlgorithms (added in crrev.com/c/6951913) and
PreferSlowCiphers (added in crrev.com/c/6986002) for the ChromeOS login
screen. These are needed because the policies affect the configuration
of TLS connections, which is required by the ChromeOS login screen
separately from the browser.

This is implemented by (in the parent CL, crrev.com/c/7037056) plumbing
the settings via profile prefs for the ChromeOS login screen Profile,
using the same pref names as we use for the local_state prefs that
ordinarily configure the corresponding user policies. In this CL, those
profile prefs are hooked up to device policies via
LoginProfilePolicyProvider.

The policies are marked experimental (future_on rather than
supported_on) because the configurations are still subject to change
during implementation and standardization.

Also adds browser tests for how these policies interact with the
base::Feature that also toggles this behavior.

Bug: 445967223, 437413977, 437415001
Change-Id: I211314476e6e42c3fc7ac86ecb712604cfe86bc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7000531
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Renato Silva <rrsilva@google.com>
Reviewed-by: Danila Kuzmin <dkuzmin@google.com>
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1545963}