[iOS] Fix home customization dismissal crash during Chromium shutdown
When the home customization sheet was open and Chromium shut down,
causing the sheet to be dismissed, the Home Customization Coordinator
would crash on stop.

This happened because SnackbarCoordinator (the dispatcher) was stopped
before the Home Customization Coordinator. As a result, it removed
itself as the protocol dispatcher while the Home Customization
Coordinator was still trying to use it, leading to a crash since
SnackbarCoordinator was no longer a valid dispatcher.

A safety check has now been added to prevent this crash.

Bug: 458060017

Change-Id: Ic5df4a1178578fb8d65fb55bb44bfbdaf4ad37be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7161575
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Commit-Queue: Pier-Alexandre Bouchard <pabouchard@google.com>
Cr-Commit-Position: refs/heads/main@{#1545984}