Introduce CCAM::Observer::OnCreditCardAccessManagerDestroyed
The original plan was to listen to AutofillDriverFactory events to
determine whether to stop observing a CCAM. However, as we will want
to emit CCAM::Observer events in CCAM's destructor, this will not work.
As a result, this CL adds an observer event for CCAM destruction,
which allows observers to safely remove themselves.

Bug: 460035068
Change-Id: I559034df5023c3f07ce04239a7f14b92a595c97a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7159090
Reviewed-by: Vinny Persky <vinnypersky@google.com>
Auto-Submit: Jan Keitel <jkeitel@google.com>
Commit-Queue: Vinny Persky <vinnypersky@google.com>
Cr-Commit-Position: refs/heads/main@{#1546003}