Report issue to DevTools for History Manipulation Intervention
Previously, there existed no user-visible indicators for when the
history manipulation intervention was triggered. This change adds
an explainer to the DevTools "issues" panel, and is accompanied
with a companion DevTools frontend CL, which is responsible for
the frontend rendering.

This CL adds the NavigationEntryMarkedSkippable Generic Issue to
Audits.pdl and needs to land first.

Devtools FE rendering companion CL:
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7118587

Bug: 445481919
Change-Id: I610cad70d8c9b0ace59e6d0334061a911b2894b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7113678
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Andrew Verge <averge@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Enoch Chigbu <enochchigbu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1546035}