Fix layout of infobar when exiting immersive full screen
During the layout of the infobar we were calling GetExtraInfobarOffset
of the ImmersiveModeController instead of the BrowserViewLayoutDelegate.
The latter checks if we are in enabled state before it returns a
non-zero offset. Otherwise when we exit immersive full screen the
infobar is placed in the wrong bounds.

Bug: 453717426
Change-Id: I3692c4ad17863069c3f44563dd935fdee2a4c2b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7155726
Reviewed-by: Darryl James <dljames@chromium.org>
Commit-Queue: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1546044}