History: Remove lazy loading, not worth it as currently used.
lazy_load.js essentially only contained synced_device_manager.js and the
final produced files were as follows (measured on Linux):

380K ./history.rollup.js
4.5K ./lazy_load.rollup.js
149K ./shared.rollup.js

Note that lazy_load.rollup.js amounts to a tiny percentage (~0.8%) of
the other two files which are loaded on startup, and therefore not worth it.

Removing it in preparation of turning on the new
@webui-eslint/web-component-missing-deps check for this folder (which
does not currently support folders using lazy loading).

Bug: 457866803
Change-Id: Id820228d8c476ddcd3b6ca5c534fb8692d7d0e47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7155715
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1546047}