[Android][Settings] Hide title breadcrumb
Search bar and the title breadcrumb are overlapped in the toolbar.
Hides the title for now, with a plan to relocate it to the top of
the content pane.

Also adds another TODO about title update in conjunction with search
results fragment.

Bug: 444468463
Change-Id: I25b4a72a599afe138e24dfa8f777e8e5ba4739b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7163335
Reviewed-by: Moe Adel <adelm@google.com>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Moe Adel <adelm@google.com>
Cr-Commit-Position: refs/heads/main@{#1546085}