[bedrock] Migrate enterprise_signin_service.cc away from BrowserListObserver.
Note that this changes the existing observer from listening to all
browsers globally to just listening to browsers within the relevant
Profile.

This migration is part of project bedrock to reduce the dependencies on
Browser and BrowserList. See https://crbug.com/431671320 for more info.

Bug: 459505408
Change-Id: I45daca00f5eea676104fff5c99b478f06a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7138686
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Commit-Queue: Glenn Hartmann <hartmanng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1546124}