fwupd: Handle unknown device ID in FirmwareUpdateManager
The device ID passed into BeginUpdate is eventually passed to OnGetFile,
which tries to match it with one of the devices in `updates_`. Crash
reports indicate that sometimes there is no match, which leads
dereferencing a null `inflight_update_`. Fix by checking for this case
and returning a new kUnknownDeviceId error.

It's likely that some additional higher-level logic fix is needed to
prevent this error from occurring in the first place, but regardless we
should have a null check here, since the callback flow is fairly
complicated.

Bug: 425162570
Change-Id: I888d3681222ea71be750c21025871c1d992ad515
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7156016
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1546140}