Do not show side panel on pin when floaty is open
When the user pins a tab from the floating UI, the side panel should
not be shown. This change adds a feature flag to control this
behavior and modifies the logic to bind the tab without showing the
side panel when the floating UI is active.

Bug: b:461567733
Change-Id: I99e68250a98a35da8cbd50d2aea5097d6a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7164120
Reviewed-by: Trevor Perrier <perrier@chromium.org>
Commit-Queue: Bryant Chandler <bryantchandler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1546173}