[GlicMI] escape hatch for WebUI crash
In case of a web ui crash we will close active embedder and shutdown webcontents. Activating an inactive embedder will reload webcontents.

I didn't find a way to first reload the webui before trying to shut it down - after the crash renderer is missing and I'm not sure if there's a good way to recover from that.

Bug: 456205343
Change-Id: I70c0ee2d89ded0c94c2b1de8f5d5157b8bb3bbbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7162374
Reviewed-by: Dan Harrington <harringtond@chromium.org>
Commit-Queue: Basia Zimirska <basiaz@google.com>
Cr-Commit-Position: refs/heads/main@{#1546217}