[layout] Add LayoutUnitDiffuser.
Occasionally in layout we'll divide a LayoutUnit by some number, (e.g.
distributing free-space with "justify-content:space-between" as one
example).

When we do this, we round down potentially leaving space unused. This
can result in a "gap" at the end of a layout the last item should be
"flush" with the end.

To compensate for this we should "diffuse" or "dither" the remainder
over the number of "buckets" we have. For example if we divided by 7 a
"nice" way to diffuse would be:

```
___X___
_X___X_
_X_X_X_
X_X_X_X
X_XXX_X
XXX_XXX
```

This patch just adds the class - followup patches will make flex/table
layout use it.

Bug: 40437604
Change-Id: Ie72927a19de3814ab3a791552a082c4fc1a6e52d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7141533
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1546247}