extensions: Convert extensionsEnabled to a static method
This is a part of the patch series to change ExtensionActionsBridge from
per-profile to per-browser.

While most of ExtensionActionsBridge methods will become per-browser,
extensionsEnabled() will continue to be per-profile for several reasons;
the decision is solely based on the profile, and it'd be a lot of work
to rewrite all its callers to pass around a browser (ChromeAndroidTask)
instead of Profile.

This patch converts ExtensionActionsBridge.extensionsEnabled() to a
static method.

However, UI code should call ExtensionUi.isEnabled() instead of
directly calling EABridge.extensionsEnabled(), then it can be faked with
FakeExtensionUiBackend. For simplicity, we drop the ability to set its
return value with FakeExtensionActionsBridge.

Bug: 461593453
Change-Id: I5be3076980cda976a2efe28feab95084d3d97f86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7165257
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1546300}