BrowsingHistoryService: Remove QueryResultsInfo::has_synced_results
This field was only used in tests, had somewhat unclear semantics (it
was not guaranteed to be updated in every request), and it was easy to
mix up with BrowsingHistoryService::has_synced_results_ which has
slightly different semantics.

Overall, doesn't seem worth the hassle to keep this.

Bug: none
Change-Id: Ie0740fcfb727509984cb30482b2ffb206a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7161974
Reviewed-by: Mahmoud Rashad <mmrashad@google.com>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1546365}